Remove aosp submodule and copy in files instead
Because repo init doesn't checkout submodules we need to copy in the
AOSP default.xml file into our git repository to be able to include it.
This way it will be a bit easier to update the whole aosp directory if
we need to pull changes from upstream.

Others copy the default.xml file and change it like:

https://github.com/LineageOS/android/blob/lineage-17.1/default.xml